requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
transformers>=4.40.0
peft>=0.10.0
torch>=2.1.0
Pillow>=10.0.0
accelerate>=0.28.0
